requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
ImageIO==2.37.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.16.3
six==1.17.0
tifffile==2025.12.20
tzdata==2025.3
